requirements.txt
fastapi==0.118.0
requests==2.32.5
uvicorn==0.37.0
pydantic==2.11.10
SQLAlchemy==2.0.43
alembic==1.16.5
python-dotenv==1.1.1
httpx==0.28.1
pydantic-settings==2.11.0
redis==6.4.0
aioredis==2.0.1
tenacity==9.1.2
celery==5.5.3
python-multipart==0.0.19
google-genai>=0.8.0
loki-logger-handler>=0.1.0
loguru>=0.7.0
